Split RendererServices into separate rendererservices.h
