+++ src/include/rendererservices.h
@@ -1,0 +1,292 @@
+/*
+Copyright (c) 2009-2014 Sony Pictures Imageworks Inc., et al.
+All Rights Reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are
+met:
+* Redistributions of source code must retain the above copyright
+  notice, this list of conditions and the following disclaimer.
+* Redistributions in binary form must reproduce the above copyright
+  notice, this list of conditions and the following disclaimer in the
+  documentation and/or other materials provided with the distribution.
+* Neither the name of Sony Pictures Imageworks nor the names of its
+  contributors may be used to endorse or promote products derived from
+  this software without specific prior written permission.
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#pragma once
+
+
+#include "oslconfig.h"
+
+#include <OpenImageIO/ustring.h>
+
+
+OSL_NAMESPACE_ENTER
+
+class RendererServices;
+struct ShaderGlobals;
+
+
+
+/// Opaque pointer to whatever the renderer uses to represent a
+/// (potentially motion-blurred) coordinate transformation.
+typedef const void * TransformationPtr;
+
+
+// Callbacks for closure creation
+typedef void (*PrepareClosureFunc)(RendererServices *, int id, void *data);
+typedef void (*SetupClosureFunc)(RendererServices *, int id, void *data);
+
+
+
+
+/// RendererServices defines an abstract interface through which a
+/// renderer may provide callback to the ShadingSystem.
+class OSLEXECPUBLIC RendererServices {
+public:
+    RendererServices () { }
+    virtual ~RendererServices () { }
+
+    /// Get the 4x4 matrix that transforms by the specified
+    /// transformation at the given time.  Return true if ok, false
+    /// on error.
+    virtual bool get_matrix (Matrix44 &result, TransformationPtr xform,
+                             float time) = 0;
+
+    /// Get the 4x4 matrix that transforms by the specified
+    /// transformation at the given time.  Return true if ok, false on
+    /// error.  The default implementation is to use get_matrix and
+    /// invert it, but a particular renderer may have a better technique
+    /// and overload the implementation.
+    virtual bool get_inverse_matrix (Matrix44 &result, TransformationPtr xform,
+                                     float time);
+
+    /// Get the 4x4 matrix that transforms by the specified
+    /// transformation.  Return true if ok, false on error.  Since no
+    /// time value is given, also return false if the transformation may
+    /// be time-varying.
+    virtual bool get_matrix (Matrix44 &result, TransformationPtr xform) = 0;
+
+    /// Get the 4x4 matrix that transforms by the specified
+    /// transformation.  Return true if ok, false on error.  Since no
+    /// time value is given, also return false if the transformation may
+    /// be time-varying.  The default implementation is to use
+    /// get_matrix and invert it, but a particular renderer may have a
+    /// better technique and overload the implementation.
+    virtual bool get_inverse_matrix (Matrix44 &result, TransformationPtr xform);
+
+    /// Get the 4x4 matrix that transforms points from the named
+    /// 'from' coordinate system to "common" space at the given time.
+    /// Returns true if ok, false if the named matrix is not known.
+    virtual bool get_matrix (Matrix44 &result, ustring from, float time) = 0;
+
+    /// Get the 4x4 matrix that transforms points from "common" space to
+    /// the named 'to' coordinate system to at the given time.  The
+    /// default implementation is to use get_matrix and invert it, but a
+    /// particular renderer may have a better technique and overload the
+    /// implementation.
+    virtual bool get_inverse_matrix (Matrix44 &result, ustring to, float time);
+
+    /// Get the 4x4 matrix that transforms 'from' to "common" space.
+    /// Since there is no time value passed, return false if the
+    /// transformation may be time-varying (as well as if it's not found
+    /// at all).
+    virtual bool get_matrix (Matrix44 &result, ustring from) = 0;
+
+    /// Get the 4x4 matrix that transforms points from "common" space to
+    /// the named 'to' coordinate system.  Since there is no time value
+    /// passed, return false if the transformation may be time-varying
+    /// (as well as if it's not found at all).  The default
+    /// implementation is to use get_matrix and invert it, but a
+    /// particular renderer may have a better technique and overload the
+    /// implementation.
+    virtual bool get_inverse_matrix (Matrix44 &result, ustring to);
+
+    /// Transform points Pin[0..npoints-1] in named coordinate system
+    /// 'from' into 'to' coordinates, storing the result in Pout[] using
+    /// the specified vector semantic (POINT, VECTOR, NORMAL).  The
+    /// function returns true if the renderer correctly transformed the
+    /// points, false if it failed (for example, because it did not know
+    /// the name of one of the coordinate systems).  A renderer is free
+    /// to not implement this, in which case the default implementation
+    /// is simply to make appropriate calls to get_matrix and
+    /// get_inverse_matrix.  The existance of this method is to allow
+    /// some renderers to provide transformations that cannot be
+    /// expressed by a 4x4 matrix.
+    ///
+    /// If npoints == 0, the function should just return true if a
+    /// known nonlinear transformation is available to transform points
+    /// between the two spaces, otherwise false.  (For this calling
+    /// pattern, sg, Pin, Pout, and time will not be used and may be 0.
+    /// As a special case, if from and to are both empty strings, it
+    /// returns true if *any* nonlinear transformations are supported,
+    /// otherwise false.
+    ///
+    /// Note to RendererServices implementations: just return 'false'
+    /// if there isn't a special nonlinear transformation between the
+    /// two spaces.
+    virtual bool transform_points (ShaderGlobals *sg,
+                                   ustring from, ustring to, float time,
+                                   const Vec3 *Pin, Vec3 *Pout, int npoints,
+                                   TypeDesc::VECSEMANTICS vectype)
+        { return false; }
+
+
+    /// Get the named attribute from the renderer and if found then
+    /// write it into 'val'.  Otherwise, return false.  If no object is
+    /// specified (object == ustring()), then the renderer should search *first*
+    /// for the attribute on the currently shaded object, and next, if
+    /// unsuccessful, on the currently shaded "scene".
+    ///
+    /// Note to renderers: if renderstate is NULL, that means
+    /// get_attribute is being called speculatively by the runtime
+    /// optimizer, and it doesn't know which object the shader will be
+    /// run on. Be robust to this situation, return 'true' (retrieve the
+    /// attribute) if you can (known object and attribute name), but
+    /// otherwise just fail by returning 'false'.
+    virtual bool get_attribute (void *renderstate, bool derivatives,
+                                ustring object, TypeDesc type, ustring name,
+                                void *val ) = 0;
+
+    /// Similar to get_attribute();  this method will return the 'index'
+    /// element of an attribute array.
+    virtual bool get_array_attribute (void *renderstate, bool derivatives,
+                                      ustring object, TypeDesc type,
+                                      ustring name, int index, void *val ) = 0;
+
+    /// Get the named user-data from the current object and write it into
+    /// 'val'. If derivatives is true, the derivatives should be written into val
+    /// as well. Return false if no user-data with the given name and type was
+    /// found.
+    virtual bool get_userdata (bool derivatives, ustring name, TypeDesc type,
+                               void *renderstate, void *val) = 0;
+
+    /// Does the current object have the named user-data associated with it?
+    virtual bool has_userdata (ustring name, TypeDesc type, void *renderstate) = 0;
+
+    /// Filtered 2D texture lookup for a single point.
+    ///
+    /// s,t are the texture coordinates; dsdx, dtdx, dsdy, and dtdy are
+    /// the differentials of s and t change in some canonical directions
+    /// x and y.  The choice of x and y are not important to the
+    /// implementation; it can be any imposed 2D coordinates, such as
+    /// pixels in screen space, adjacent samples in parameter space on a
+    /// surface, etc.
+    ///
+    /// Return true if the file is found and could be opened, otherwise
+    /// return false.
+    virtual bool texture (ustring filename, TextureOpt &options,
+                          ShaderGlobals *sg,
+                          float s, float t, float dsdx, float dtdx,
+                          float dsdy, float dtdy, float *result);
+
+    /// Filtered 3D texture lookup for a single point.
+    ///
+    /// P is the volumetric texture coordinate; dPd{x,y,z} are the
+    /// differentials of P in some canonical directions x, y, and z.
+    /// The choice of x,y,z are not important to the implementation; it
+    /// can be any imposed 3D coordinates, such as pixels in screen
+    /// space and depth along the ray, etc.
+    ///
+    /// Return true if the file is found and could be opened, otherwise
+    /// return false.
+    virtual bool texture3d (ustring filename, TextureOpt &options,
+                            ShaderGlobals *sg, const Vec3 &P,
+                            const Vec3 &dPdx, const Vec3 &dPdy,
+                            const Vec3 &dPdz, float *result);
+
+    /// Filtered environment lookup for a single point.
+    ///
+    /// R is the directional texture coordinate; dRd[xy] are the
+    /// differentials of R in canonical directions x, y.
+    ///
+    /// Return true if the file is found and could be opened, otherwise
+    /// return false.
+    virtual bool environment (ustring filename, TextureOpt &options,
+                              ShaderGlobals *sg, const Vec3 &R,
+                              const Vec3 &dRdx, const Vec3 &dRdy, float *result);
+
+    /// Get information about the given texture.  Return true if found
+    /// and the data has been put in *data.  Return false if the texture
+    /// doesn't exist, doesn't have the requested data, if the data
+    /// doesn't match the type requested. or some other failure.
+    virtual bool get_texture_info (ShaderGlobals *sg,
+                                   ustring filename, int subimage,
+                                   ustring dataname, TypeDesc datatype,
+                                   void *data);
+
+
+    /// Lookup nearest points in a point cloud. It will search for
+    /// points around the given center within the specified radius. A
+    /// list of indices is returned so the programmer can later retrieve
+    /// attributes with pointcloud_get. The indices array is mandatory,
+    /// but distances can be NULL.  If a derivs_offset > 0 is given,
+    /// derivatives will be computed for distances (when provided).
+    ///
+    /// Return the number of points found, always < max_points
+    virtual int pointcloud_search (ShaderGlobals *sg,
+                                   ustring filename, const OSL::Vec3 &center,
+                                   float radius, int max_points, bool sort,
+                                   size_t *out_indices,
+                                   float *out_distances, int derivs_offset);
+
+    /// Retrieve an attribute for an index list. The result is another array
+    /// of the requested type stored in out_data.
+    ///
+    /// Return 1 if the attribute is found, 0 otherwise.
+    virtual int pointcloud_get (ShaderGlobals *sg,
+                                ustring filename, size_t *indices, int count,
+                                ustring attr_name, TypeDesc attr_type,
+                                void *out_data);
+
+    /// Write a point to the named pointcloud, which will be saved
+    /// at the end of the frame.  Return true if everything is ok,
+    /// false if there was an error.
+    virtual bool pointcloud_write (ShaderGlobals *sg,
+                                   ustring filename, const OSL::Vec3 &pos,
+                                   int nattribs, const ustring *names,
+                                   const TypeDesc *types,
+                                   const void **data);
+
+    /// Options for the trace call.
+    struct TraceOpt {
+        float mindist;    ///< ignore hits closer than this
+        float maxdist;    ///< ignore hits farther than this
+        bool shade;       ///< whether to shade what is hit
+        ustring traceset; ///< named trace set
+        TraceOpt () : mindist(0.0f), maxdist(1.0e30), shade(false) { }
+    };
+
+    /// Immediately trace a ray from P in the direction R.  Return true
+    /// if anything hit, otherwise false.
+    virtual bool trace (TraceOpt &options, ShaderGlobals *sg,
+                        const OSL::Vec3 &P, const OSL::Vec3 &dPdx,
+                        const OSL::Vec3 &dPdy, const OSL::Vec3 &R,
+                        const OSL::Vec3 &dRdx, const OSL::Vec3 &dRdy) {
+        return false;
+    }
+
+    /// Get the named message from the renderer and if found then
+    /// write it into 'val'.  Otherwise, return false.  This is only
+    /// called for "sourced" messages, not ordinary intra-group messages.
+    virtual bool getmessage (ShaderGlobals *sg, ustring source, ustring name,
+                             TypeDesc type, void *val, bool derivatives) {
+        return false;
+    }
+};
+
+
+OSL_NAMESPACE_EXIT
